backend/requirements.txt
fastapi
fastapi-mail
uvicorn
motor
python-dotenv
pydantic
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
